+++ frontend/src/components/MyDrawer.js
@@ -4,8 +4,10 @@ import {
   Box,
   List,
   ListItemButton,
-  ListItemText
+  ListItemText,
+  Typography
 } from '@mui/material';
+import Divider from '@mui/material/Divider';
 
 export default function MyDrawer({ open, setDrawerOpen }) {
   return (
@@ -13,17 +15,30 @@ export default function MyDrawer({ open, setDrawerOpen }) {
       open={open}
       onClose={() => setDrawerOpen(false)}
       anchor="right"
+      sx={{
+        '& .MuiDrawer-paper': {
+          bgcolor: 'var(--secondary-background-color)',
+          color: 'var(--secondary-color)',
+          borderTopLeftRadius: 15,
+          borderBottomLeftRadius: 15
+        },
+        '& .MuiListItemText-primary': {
+          fontWeight: 'bold'
+        },
+      }}
     >
-      <Toolbar /> {/* pushes content below AppBar */}
+      <Typography variant="h4" align="center" gutterBottom sx={{ mt: 2, mx: 1.5, fontWeight: 'bold', color: 'var(--secondary-color)'}}>
+        notoli 
+      </Typography>
       <Box role="navigation">
-        <List>
+        <Divider sx={{ borderBottomWidth: 2, mx: 1, my: 0.1, bgcolor: 'var(--secondary-color)' }} />
+        {/* disablePadding + my: 0 is helping reduce the padding, but not making it smaller like I want */}
+        <List disablePadding sx={{ my: 0 }}>
           <ListItemButton>
-            <ListItemText primary="Workspaces" />
-          </ListItemButton>
-          <ListItemButton>
-            <ListItemText primary="Settings" />
+            <ListItemText primary="Workspace" secondary="This is a Workspace"/> 
           </ListItemButton>
         </List>
+        <Divider sx={{ borderBottomWidth: 2, mx: 1, my: 0.1, bgcolor: 'var(--secondary-color)' }} />
       </Box>
     </Drawer>
   );